% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{build_discrete_legend}
\alias{build_discrete_legend}
\title{Build discrete legend}
\usage{
build_discrete_legend(data, fill_col, label_col)
}
\arguments{
\item{data}{Atlas data}

\item{fill_col}{Name of the fill column}

\item{label_col}{Name of the label column}
}
\value{
List with discrete legend specification or NULL if too many
  categories
}
\description{
Creates a discrete legend specification for categorical data.
}
\keyword{internal}
