% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain-mesh.R
\name{compute_parallel_transp_fr}
\alias{compute_parallel_transp_fr}
\title{Compute parallel transport frames along curve}
\usage{
compute_parallel_transp_fr(curve)
}
\description{
Compute parallel transport frames along curve
}
\keyword{internal}
