% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{create_ggseg3d_widget}
\alias{create_ggseg3d_widget}
\title{Create ggseg3d htmlwidget}
\usage{
create_ggseg3d_widget(meshes, legend_data)
}
\arguments{
\item{meshes}{List of mesh data structures}

\item{legend_data}{Legend specification (or NULL)}
}
\value{
An htmlwidget object of class "ggseg3d"
}
\description{
Creates the final htmlwidget object with default options and sizing policy.
}
\keyword{internal}
