% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-types.R
\name{data_merge_mesh}
\alias{data_merge_mesh}
\title{Merge user data with mesh atlas data}
\usage{
data_merge_mesh(.data, atlas_data)
}
\arguments{
\item{.data}{User-provided data frame}

\item{atlas_data}{Atlas data frame}
}
\value{
Merged data frame
}
\description{
Merge user data with mesh atlas data
}
\keyword{internal}
