% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg3d_atlas.R
\name{ggseg3d_atlas-class}
\alias{ggseg3d_atlas-class}
\alias{as_ggseg3d_atlas}
\alias{ggseg3d_atlas}
\title{`ggseg3d_atlas` class}
\usage{
as_ggseg3d_atlas(x, return = FALSE)
}
\arguments{
\item{x}{data.frame to be made a ggseg-atlas}

\item{return}{return logical}
}
\value{
an object of class 'ggseg3d_atlas'. A nested tibble of different
   brain surface shapes, hemispheres and tri-surface mesh information
   for different brain regions in a specific atlas.
}
\description{
The `ggseg_3datlas` class is a subclass of [`data.frame`][base::data.frame()],
created in order to have different default behaviour. It heavily relies on
the "tibble" [`tbl_df`][tibble::tibble()].
[tidyverse](https://www.tidyverse.org/packages/), including
[dplyr](http://dplyr.tidyverse.org/),
[ggplot2](http://ggplot2.tidyverse.org/),
[tidyr](http://tidyr.tidyverse.org/), and
[readr](http://readr.tidyverse.org/).
}
\section{Properties of `ggseg3d_atlas`}{


Objects of class `ggseg3d_atlas` have:
* A `class` attribute of `c("ggseg3d_atlas", "tbl_df", "tbl", "data.frame")`.
* A base type of `"list"`, where each element of the list has the same
  [NROW()].
* A lot of this script and its functions are taken from the
  [`tibble`][tibble::tibble()]-package
}

\examples{
tmp <- as.data.frame(dk_3d)
class(tmp)
new_atlas <- as_ggseg3d_atlas(tmp)
class(new_atlas)
}
\seealso{
[tibble()], [as_tibble()], [tribble()], [print.tbl()], [glimpse()]
}
