% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-types.R
\name{is_tract_atlas}
\alias{is_tract_atlas}
\title{Check if atlas is a tract atlas}
\usage{
is_tract_atlas(atlas)
}
\arguments{
\item{atlas}{An atlas object to check}
}
\value{
Logical indicating if this is a tract atlas
}
\description{
Check if atlas is a tract atlas
}
