% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes.R
\name{remove_axes}
\alias{remove_axes}
\title{Remove axis information from ggseg3d plot}
\usage{
remove_axes(p)
}
\arguments{
\item{p}{plotly object}
}
\value{
plotly object without axes
}
\description{
When publishing data visualisation in 3d mesh plots
in general the axes are not important, atleast
they are not for ggseg3d, where the axis values
are arbitrary.
}
\examples{
library(magrittr)
ggseg3d() \%>\%
   remove_axes()
}
