% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions.R
\name{set_background}
\alias{set_background}
\title{Set background color of ggseg3d plot}
\usage{
set_background(p, colour = "#ffffff")
}
\arguments{
\item{p}{A \code{ggseg3d} widget or \code{ggsegray} rgl object.}

\item{colour}{string. Background color (hex or named color)}
}
\value{
The input object (modified), for piping.
}
\description{
Changes the background color of a ggseg3d widget or ggsegray rgl scene.
}
\examples{
ggseg3d() |> set_background("black")
\dontrun{
# rgl requires OpenGL; not run in check environments.
ggsegray(atlas = dk()) |> set_background("black")
}
}
\seealso{
Other brain modifiers:
\code{\link[=add_glassbrain]{add_glassbrain()}},
\code{\link[=pan_camera]{pan_camera()}},
\code{\link[=set_dimensions]{set_dimensions()}},
\code{\link[=set_edges]{set_edges()}},
\code{\link[=set_flat_shading]{set_flat_shading()}},
\code{\link[=set_legend]{set_legend()}},
\code{\link[=set_orthographic]{set_orthographic()}},
\code{\link[=set_positioning]{set_positioning()}}
}
\concept{brain modifiers}
