% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-output.R
\name{updateGgseg3dBackground}
\alias{updateGgseg3dBackground}
\title{Update background in Shiny}
\usage{
updateGgseg3dBackground(session, outputId, colour)
}
\arguments{
\item{session}{The Shiny session object}

\item{outputId}{The output ID of the ggseg3d widget}

\item{colour}{Background color (hex or named color)}
}
\description{
Sends a message to update the background color of a ggseg3d widget
in a Shiny app.
}
