% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-output.R
\name{updateGgseg3dCamera}
\alias{updateGgseg3dCamera}
\title{Update camera in Shiny}
\usage{
updateGgseg3dCamera(session, outputId, camera)
}
\arguments{
\item{session}{The Shiny session object}

\item{outputId}{The output ID of the ggseg3d widget}

\item{camera}{Camera position preset or custom position}
}
\description{
Sends a message to update the camera position of a ggseg3d widget
in a Shiny app.
}
