% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg3d-package.R
\docType{package}
\name{ggseg3d-package}
\alias{ggseg3d-package}
\title{ggseg3d: Plot brain segmentations in 3D}
\description{
Plots statistics on interactive 3D brain meshes, rendered with Three.js
through htmlwidgets. Atlases are \code{ggseg_atlas} objects from
\code{ggseg.formats}; \code{dk}, \code{aseg} and \code{tracula} are re-exported here for
convenience.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggseg3d}
  \item \url{https://ggsegverse.github.io/ggseg3d/}
  \item Report bugs at \url{https://github.com/ggsegverse/ggseg3d/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})
  \item Didac Vidal-Piñeiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

Other contributors:
\itemize{
  \item Center for Lifespan Changes in Brain and Cognition (LCBC), University of Oslo [copyright holder]
  \item three.js authors (Three.js and OrbitControls libraries, see LICENSE.note) [contributor, copyright holder]
}

}
\keyword{internal}
