% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg3d-package.R
\docType{package}
\name{ggseg3d-package}
\alias{ggseg3d-package}
\title{ggseg3d: Plot brain segmentations in 3D}
\description{
Plotting results from analyses done on data derived from brain
segmentations is a common need, but may be quite laborious.
Results from such analyses are usually easier to interpret if
the plot can mimmick the shape and position in the brain it
represents.
}
\details{
This package contains data from various brain parcellations,
with convenient functions to inspect the results directly on
a brain-plot using interactive 3D mesh visualizations powered
by Three.js via htmlwidgets.

The package uses `ggseg_atlas` objects from ggseg.formats
that contain 3D vertex mappings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggseg3d}
  \item \url{https://ggsegverse.github.io/ggseg3d/}
  \item Report bugs at \url{https://github.com/ggsegverse/ggseg3d/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Piñeiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
