% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg3d-package.R, R/ggseg3d.R
\docType{package}
\name{ggseg3d}
\alias{ggseg3d}
\alias{ggseg3d-package}
\title{ggseg3d: Plot brain segmentations with plotly}
\usage{
ggseg3d(.data = NULL, atlas = "dkt_3d", surface = "LCBC",
  hemisphere = c("right", "subcort"), label = "area", text = NULL,
  colour = "colour", palette = NULL, na.colour = "darkgrey",
  na.alpha = 1, show.legend = TRUE)
}
\arguments{
\item{.data}{A data.frame to use for plot aesthetics. Must include a
column called "area" corresponding to areas.}

\item{atlas}{Either a string with the name of a 3d atlas to use.}

\item{surface}{String. Which surface to plot. Either "pial","white", or "inflated"[default]}

\item{hemisphere}{String. Hemisphere to plot. Either "left" or "right"[default],
can also be "subcort".}

\item{label}{String. Quoted name of column in atlas/data that should be used to name traces}

\item{text}{String. Quoated name of column in atlas/data that should be added as extra
information in the hover text.}

\item{colour}{String. Quoted name of column from which colour should be supplied}

\item{palette}{String. Vector of colour names or HEX colours. Can also be a named
numeric vector, with colours as names, and breakpoint for that colour as the value}

\item{na.colour}{String. Either name, hex of RGB for colour of NA in colour.}

\item{na.alpha}{Numeric. A number between 0 and 1 to control transparency of NA-regions.}

\item{show.legend}{Logical. Toggle legend if colour is numeric.}
}
\value{
a plotly object
}
\description{
Plotting results from analyses done on data derived from brain
segmentations is a common need, but may be quite laborious.
Results from such analyses are usually easier to interpret if
the plot can mimmick the shape and position in the brain it
represents.

\code{ggseg3d} plots and returns a plotly mesh3d object.
}
\details{
This package contains data from various brain parcellations,
with convenient functions to inspect the results directly on
a brain-plot wither with plotly in 3d meshes.

\describe{
\strong{Available atlases:}
  \item{`dkt_3d`}{
    The Desikan-Killiany Cortical Atlas [default], Freesurfer cortical segmentations, in 3dmesh format
  }
}

\strong{Available surfaces:}
\itemize{
\item `inflated` Fully inflated surface
\item `white` white matter surface
 }
}
\examples{
ggseg3d()
ggseg3d(surface="white")
ggseg3d(surface="inflated")
ggseg3d(show.legend = FALSE)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LCBC-UiO/ggseg3d}
  \item Report bugs at \url{https://github.com/LCBC-UiO/ggseg3d/issues}
}


\code{\link[plotly]{plot_ly}}, \code{\link[plotly]{add_trace}}, \code{\link[plotly]{layout}}, the plotly package
}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (0000-0002-5756-0223)

Authors:
\itemize{
  \item Didac Vidal Pineiro \email{d.v.pineiro@psykologi.uio.no} (0000-0001-9997-9156)
}


Athanasia Mowinckel and Didac Pineiro
}
\keyword{internal}
