% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera.R
\name{pan_camera}
\alias{pan_camera}
\title{Pan camera position of ggseg3d plot}
\usage{
pan_camera(p, camera)
}
\arguments{
\item{p}{plotly object}

\item{camera}{string or list.}
}
\value{
plotly object
}
\description{
The default position for plotly
mesh plots are not satifsying for
brain plots. This convenience function
can pan the camera to lateral or medial
view, or to custom made views if you are
plotly savvy.
}
\examples{
library(magrittr)
ggseg3d() \%>\%
   pan_camera("lateral")
}
