% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions.R
\name{pan_camera}
\alias{pan_camera}
\title{Pan camera position of ggseg3d plot}
\usage{
pan_camera(p, camera)
}
\arguments{
\item{p}{ggseg3d widget object}

\item{camera}{string or list. Camera position preset name or custom eye
  position.

\strong{Available camera presets:}
\itemize{
\item `left lateral` or `left_lateral`
\item `left medial` or `left_medial`
\item `right lateral` or `right_lateral`
\item `right medial` or `right_medial`
\item `left superior` or `left_superior`
\item `right superior` or `right_superior`
\item `left inferior` or `left_inferior`
\item `right inferior` or `right_inferior`
\item `left anterior` or `left_anterior`
\item `right anterior` or `right_anterior`
\item `left posterior` or `left_posterior`
\item `right posterior` or `right_posterior`
}}
}
\value{
ggseg3d widget object with updated camera
}
\description{
Sets the camera position for a ggseg3d widget
to standard anatomical views or custom positions.
}
\examples{
\dontrun{
ggseg3d() |>
  pan_camera("right lateral")
}
}
