% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions.R
\name{set_background}
\alias{set_background}
\title{Set background color of ggseg3d plot}
\usage{
set_background(p, colour = "#ffffff")
}
\arguments{
\item{p}{ggseg3d widget object}

\item{colour}{string. Background color (hex or named color)}
}
\value{
ggseg3d widget object with updated background
}
\description{
Changes the background color of a ggseg3d widget.
}
\examples{
\dontrun{
ggseg3d() |>
  set_background("black")
}
}
