% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions.R
\name{set_background}
\alias{set_background}
\title{Set background color of ggseg3d plot}
\usage{
set_background(p, colour = "#ffffff")
}
\arguments{
\item{p}{A `ggseg3d` widget or `ggsegray` rgl object.}

\item{colour}{string. Background color (hex or named color)}
}
\value{
The input object (modified), for piping.
}
\description{
Changes the background color of a ggseg3d widget or ggsegray rgl scene.
}
\examples{
\dontrun{
ggseg3d() |> set_background("black")

ggsegray(atlas = dk()) |> set_background("black")
}
}
