% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions.R
\name{set_dimensions}
\alias{set_dimensions}
\title{Set widget dimensions}
\usage{
set_dimensions(p, width = NULL, height = NULL)
}
\arguments{
\item{p}{ggseg3d widget object}

\item{width}{numeric. Widget width in pixels (NULL for default)}

\item{height}{numeric. Widget height in pixels (NULL for default)}
}
\value{
ggseg3d widget object with updated dimensions
}
\description{
Changes the width and height of a ggseg3d widget.
}
\examples{
\dontrun{
ggseg3d() |>
  set_dimensions(width = 800, height = 600)
}
}
