% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions.R
\name{set_legend}
\alias{set_legend}
\title{Set legend visibility}
\usage{
set_legend(p, show = TRUE)
}
\arguments{
\item{p}{ggseg3d widget object}

\item{show}{logical. Whether to show the legend (default: TRUE)}
}
\value{
ggseg3d widget object with updated legend visibility
}
\description{
Controls visibility of the colour bar legend for numeric data.
}
\examples{
\dontrun{
ggseg3d() |> set_legend(FALSE)
}
}
